shadow 4.14.0-rc2

 Conflicts:
	shadow/configure.ac
	  f88b2b7eb06f76 "Merge tag 'v4.12.3' into wip-4.12"
	  4a20d8ddd41d96 "add setlocale to AC_CHECK_FUNCS list"
	    Patch updated.

	shadow/lib/fields.c
	  31f64d99c51323 "lib: Add SAFE_PWDB_FIELDS variable"
	  6eba7fc67a1998 "valid_field: Always reject control characters"
	    Patch updated.

	shadow/lib/getdef.c
	  a8d24f06d139fa "apply and update shadow-4.0.4.1-owl-alt-usergroupname_max.patch"
	  7a7c6a57de4549 "apply and update shadow-4.0.4.1-owl-alt-crypt_gensalt.patch"
	    Resolved conflict.

	shadow/lib/prototypes.h
	  4214d2e90cd0ee "apply and update shadow-4.0.4.1-alt-progname.patch"
	    Resolved conflict.
	  dc89acf3e8e5f7 "tcb: Add remove_tcbdir() function"
	    Resolved conflict.

	shadow/libmisc/Makefile.am
	  4214d2e90cd0ee "apply and update shadow-4.0.4.1-alt-progname.patch"
	    Resolved conflict.

	shadow/libmisc/basename.c
	  Removed by us.

	shadow/libmisc/chkname.c
	  a8d24f06d139fa "apply and update shadow-4.0.4.1-owl-alt-usergroupname_max.patch"
	    Patch updated (see upstream commit 3b7cc053872cf4ce77fd74dc037f43f34e951e83
	      "lib: replace `USER_NAME_MAX_LENGTH` macro").

	shadow/libmisc/copydir.c
	  f88b2b7eb06f76 "Merge tag 'v4.12.3' into wip-4.12"
	  5642300d19265a "apply and update shadow-4.0.4.1-alt-copy_tree-perms.patch"
	    Resolved conflict.

	shadow/libmisc/find_new_gid.c
	  2785f44019ef55 "Merge tag 'v4.5'"
	    Patch updated (default SYS_GID_MIN).

	shadow/libmisc/find_new_uid.c
	  2785f44019ef55 "Merge tag 'v4.5'"
	    Patch updated (default SYS_UID_MIN).

	shadow/libmisc/pam_pass_non_interactive.c
	  b70bcb2a0334bc "Merge ALT-specific tcb patch"
	    Patch updated.

	shadow/libmisc/prefix_flag.c
	  1a5cef2c1abb13 "Ensure that prefix is not '/'"
	    Resloved conflict.
	  f18da56c783c73 "Add prefix support for TCB"
	    Patch updated.

	shadow/libmisc/salt.c
	  7a7c6a57de4549 "apply and update shadow-4.0.4.1-owl-alt-crypt_gensalt.patch"
	    Resloved conflict.

	shadow/libmisc/utmp.c
	  c8032c33cf1bdf "apply shadow-4.1.4.2-fixes.patch"
	    Patch updated.

	shadow/man/Makefile.am
	  4961478a60978e "apply and update shadow-4.0.4.1-alt-makefile.patch"
	    Resloved conflict.

	shadow/src/Makefile.am
	  6f899fce28f471 "Move groupmems to /usr/bin/ from /usr/sbin/"
	    Patch updated.
	  e54d80c0839180 "fixed build without TCB and/or PAM"
	    Patch updated.
	  4961478a60978e "apply and update shadow-4.0.4.1-alt-makefile.patch"
	    Resloved conflict.

	shadow/src/chpasswd.c
	  b70bcb2a0334bc "Merge ALT-specific tcb patch"
	    Patch updated.

	shadow/src/chsh.c
	  5115084aa563b9 "chsh: More strict checks for shell."
	  f499c654e716c4 "chsh: Fix duplicate warning"
	    Resloved conflict.

	shadow/src/groups.c
	  4214d2e90cd0ee "apply and update shadow-4.0.4.1-alt-progname.patch"
	    Resloved conflict.

	shadow/src/newgrp.c
	  4214d2e90cd0ee "apply and update shadow-4.0.4.1-alt-progname.patch"
	    Patch updated (see upstream commit 9df4801e0b65073cc8a9031b22a73532ef7fdc2c
	      "newgrp: fix potential string injection").

	shadow/src/newusers.c
	  b70bcb2a0334bc "Merge ALT-specific tcb patch"
	    Patch updated (see upstream commit 5b117d55261cd4882df9571d2801dc4b1da2176f
	      "newusers: Fail early").

	shadow/src/useradd.c
	  7a92706f211236 "useradd: Set default group to 100 (users)"
	    Resolved conflict.
	  b70bcb2a0334bc "Merge ALT-specific tcb patch"
	    Resolved conflict.
	  11ee5d7559a2a0 "Merge tag 'v4.8.1'"
	  f11ed4a75a70cf "Merge tag 'v4.6'"
	  b1f7e9de1abe21 "useradd: Check chown/chmod return status"
	    Patch updated.
	  20bb0027d6e67c "useradd: Add 'private' to allowed values of CREATE_MAIL_SPOOL."
	    Patch updated.
	  d51d8414fcdbe4 "useradd: Fix skel files SELinux context"
	    Patch updated (see upstream commit c80788a3ac092bc5abfa89ff48060d3f95cd5812
	      "useradd/usermod: add --selinux-range argument").

	shadow/src/userdel.c
	  77aa732ba68424 "userdel: Fix mailbox removing"
	    Resloved conflict.
	  dc89acf3e8e5f7 "tcb: Add remove_tcbdir() function"
	    Resloved conflict: remove_tcbdir() moved to libmisc/remove_tcbdir.c.

	shadow/src/usermod.c
	  1c59aaca08fb7c "apply shadow-4.0.4.1-owl-usermod-unlock.patch"
	    Resloved conflict.

+++ shadow/lib/tcbfuncs.c
@@ -147,7 +147,7 @@ static /*@null@*/ char *shadowtcb_path_rel_existing (const char *prefix_dir, con
 		         shadow_progname, link);
 		return NULL;
 	}
-	link[(size_t)ret] = '\0';
+	link[ret] = '\0';
 	rval = strdup (link);
 	if (NULL == rval) {
 		OUT_OF_MEMORY;
